requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-model-utils==4.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.8.4
dj-database-url=0.5.0
